test->main : passagem do cógido de test para main
